Explicitly get the latest work state when updating heartbeat
The previous update_heartbeat method assumed that, if there was
a worker update within the update_freq for a given host, there
would only be a single instance. However, we have seen cases in the
wild where this was not the case. This code guards against this
edge-case by using .latest() rather than .get().